Question_Parser Script Added
Question_Parser will read .csv files with correct formatting into a question bank and handles different types of questions.
So far it handles questions with up to 4 variables and calculates correct and incorrect answers. All answers for these type of questions must be written as equations that cannot have the same answers with the same values.
It will also handle questions without variables and any type of answers.